backend/requirements.txt
anyio==4.12.0
arrow==1.3.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
binaryornot==0.4.4
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cookiecutter==2.6.0
cryptography==45.0.6
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.1
django-allauth==65.11.0
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-tailwind==4.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
firebase_admin==7.1.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.9.0
zope.interface==7.2
